requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-otp==1.1.4
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.76.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
Pillow==9.4.0
protobuf==4.21.12
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfcm==1.5.4
pyparsing==3.0.9
python-http-client==3.3.7
pytz==2022.7
requests==2.28.2
rsa==4.9
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0